requirements.txt
Django==4.2.7 
djangorestframework==3.14.0 
django-cors-headers==4.3.1 
psycopg2-binary==2.9.7 
pandas==2.1.3 
scikit-learn==1.3.2 
requests==2.31.0 
